requirements.txt
doclayout_yolo==0.0.3
Flask
numpy
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
paddleocr==2.9.1
paddleocr.egg==info
Pillow
torch==2.2.2+cu118
waitress==3.0.2
huggingface_hub
